judy, Update:: 콤보박스 메뉴 초기값  null로 수정, 페이지 뒤로가기 시 드롭다운 제거, icon 데이터 타입 Widget에서 Icon으로 변경 및  변수명 다수 변경 #65
